feature/create_bootcamp_endpoint-add-mongoose: Update module and structure
