+++ src/app.module.ts
@@ -1,9 +1,10 @@
-import { BootcampsModule } from './controllers/bootcamps/bootcamps.module';
 import { Module } from '@nestjs/common';
-import { AppController } from './app.controller';
-import { AppService } from './app.service';
 import { MongooseModule } from '@nestjs/mongoose';
 import { ConfigModule } from '@nestjs/config';
+
+import { AppController } from './app.controller';
+import { AppService } from './app.service';
+import { BootcampsModule } from './modules/bootcamps.module';
 
 const mongoUrl =
   process.env.NODE_ENV === 'production'
